+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/upgrade/UpgradeRegistry.java
@@ -1,16 +1,15 @@
 package com.refinedmods.refinedstorage2.platform.api.upgrade;
 
 import java.util.Set;
-import java.util.function.Supplier;
 
 import net.minecraft.world.item.Item;
 import org.apiguardian.api.API;
 
 @API(status = API.Status.STABLE, since = "2.0.0-milestone.2.2")
 public interface UpgradeRegistry {
-    void addApplicableUpgrade(UpgradeDestination destination, Supplier<Item> itemSupplier, int maxAmount);
+    void add(UpgradeDestination destination, Item upgradeItem, int maxAmount);
 
-    Set<ApplicableUpgrade> getApplicableUpgrades(UpgradeDestination destination);
+    Set<UpgradeMapping> getByDestination(UpgradeDestination destination);
 
-    Set<UpgradeInDestination> getDestinations(Item item);
+    Set<UpgradeMapping> getByUpgradeItem(Item upgradeItem);
 }