+++ refinedstorage2-platform-api/src/main/java/com/refinedmods/refinedstorage2/platform/api/upgrade/UpgradeState.java
@@ -5,5 +5,5 @@ import org.apiguardian.api.API;
 @API(status = API.Status.STABLE, since = "2.0.0-milestone.2.10")
 @FunctionalInterface
 public interface UpgradeState {
-    boolean hasUpgrade(UpgradeItem item);
+    boolean hasUpgrade(UpgradeItem upgradeItem);
 }