+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/AbstractModInitializer.java
@@ -91,60 +91,60 @@ public abstract class AbstractModInitializer {
         );
     }
 
-    protected void addApplicableUpgrades() {
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+    protected void addUpgradeMappings() {
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.IMPORTER,
-            Items.INSTANCE::getSpeedUpgrade,
+            Items.INSTANCE.getSpeedUpgrade(),
             4
         );
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.IMPORTER,
-            Items.INSTANCE::getStackUpgrade,
+            Items.INSTANCE.getStackUpgrade(),
             1
         );
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.EXPORTER,
-            Items.INSTANCE::getSpeedUpgrade,
+            Items.INSTANCE.getSpeedUpgrade(),
             4
         );
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.EXPORTER,
-            Items.INSTANCE::getStackUpgrade,
+            Items.INSTANCE.getStackUpgrade(),
             1
         );
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.DESTRUCTOR,
-            Items.INSTANCE::getSpeedUpgrade,
+            Items.INSTANCE.getSpeedUpgrade(),
             4
         );
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.DESTRUCTOR,
-            Items.INSTANCE::getFortune1Upgrade,
+            Items.INSTANCE.getFortune1Upgrade(),
             1
         );
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.DESTRUCTOR,
-            Items.INSTANCE::getFortune2Upgrade,
+            Items.INSTANCE.getFortune2Upgrade(),
             1
         );
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.DESTRUCTOR,
-            Items.INSTANCE::getFortune3Upgrade,
+            Items.INSTANCE.getFortune3Upgrade(),
             1
         );
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.DESTRUCTOR,
-            Items.INSTANCE::getSilkTouchUpgrade,
+            Items.INSTANCE.getSilkTouchUpgrade(),
             1
         );
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.CONSTRUCTOR,
-            Items.INSTANCE::getSpeedUpgrade,
+            Items.INSTANCE.getSpeedUpgrade(),
             4
         );
-        PlatformApi.INSTANCE.getUpgradeRegistry().addApplicableUpgrade(
+        PlatformApi.INSTANCE.getUpgradeRegistry().add(
             UpgradeDestinations.CONSTRUCTOR,
-            Items.INSTANCE::getStackUpgrade,
+            Items.INSTANCE.getStackUpgrade(),
             1
         );
     }