+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/internal/upgrade/UpgradeRegistryImpl.java
@@ -1,8 +1,7 @@
 package com.refinedmods.refinedstorage2.platform.common.internal.upgrade;
 
-import com.refinedmods.refinedstorage2.platform.api.upgrade.ApplicableUpgrade;
 import com.refinedmods.refinedstorage2.platform.api.upgrade.UpgradeDestination;
-import com.refinedmods.refinedstorage2.platform.api.upgrade.UpgradeInDestination;
+import com.refinedmods.refinedstorage2.platform.api.upgrade.UpgradeMapping;
 import com.refinedmods.refinedstorage2.platform.api.upgrade.UpgradeRegistry;
 
 import java.util.Collections;
@@ -10,37 +9,44 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
-import java.util.function.Supplier;
 
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
 
 public class UpgradeRegistryImpl implements UpgradeRegistry {
-    private final Map<UpgradeDestination, Set<ApplicableUpgrade>> destinationMap = new HashMap<>();
+    private final Map<UpgradeDestination, Set<UpgradeMapping>> byDestination = new HashMap<>();
+    private final Map<Item, Set<UpgradeMapping>> byUpgradeItem = new HashMap<>();
 
     @Override
-    public void addApplicableUpgrade(final UpgradeDestination destination,
-                                     final Supplier<Item> itemSupplier,
-                                     final int maxAmount) {
-        destinationMap.computeIfAbsent(destination, key -> new HashSet<>())
-            .add(new ApplicableUpgrade(itemSupplier, maxAmount));
+    public void add(final UpgradeDestination destination, final Item upgradeItem, final int maxAmount) {
+        final ItemStack displayItemStack = new ItemStack(upgradeItem);
+        final UpgradeMapping mapping = new UpgradeMapping(
+            destination,
+            upgradeItem,
+            maxAmount,
+            upgradeItem.getName(displayItemStack).copy()
+                .append(" ")
+                .append("(")
+                .append(String.valueOf(maxAmount))
+                .append(")"),
+            destination.getName().copy()
+                .append(" ")
+                .append("(")
+                .append(String.valueOf(maxAmount))
+                .append(")"),
+            displayItemStack
+        );
+        byDestination.computeIfAbsent(destination, key -> new HashSet<>()).add(mapping);
+        byUpgradeItem.computeIfAbsent(upgradeItem, key -> new HashSet<>()).add(mapping);
     }
 
     @Override
-    public Set<ApplicableUpgrade> getApplicableUpgrades(final UpgradeDestination destination) {
-        return destinationMap.getOrDefault(destination, Collections.emptySet());
+    public Set<UpgradeMapping> getByDestination(final UpgradeDestination destination) {
+        return byDestination.getOrDefault(destination, Collections.emptySet());
     }
 
     @Override
-    public Set<UpgradeInDestination> getDestinations(final Item item) {
-        final Set<UpgradeInDestination> result = new HashSet<>();
-        for (final var entry : destinationMap.entrySet()) {
-            final UpgradeDestination destination = entry.getKey();
-            for (final ApplicableUpgrade applicableUpgrade : entry.getValue()) {
-                if (applicableUpgrade.itemSupplier().get() == item) {
-                    result.add(new UpgradeInDestination(destination, applicableUpgrade.maxAmount()));
-                }
-            }
-        }
-        return result;
+    public Set<UpgradeMapping> getByUpgradeItem(final Item upgradeItem) {
+        return byUpgradeItem.getOrDefault(upgradeItem, Collections.emptySet());
     }
 }