+++ CHANGELOG.md
@@ -14,6 +14,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Grid slot hints that show which resource will be inserted or extracted in a Grid.
 -   Help information to the side buttons by pressing SHIFT.
 -   A warning to the "filter mode" button on the storage screens if there is an allowlist with no configured filters.
+-   The "supported by" tooltip on upgrade items now shows the devices that accept the upgrade.
 
 ### Changed